LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/overlapping-navigations-and-traversals/cross-document-traversal-same-document-traversal-expected.txt


FAIL traversals in the same (back) direction: coalesced assert_equals: first load event must be going back (hash) expected "" but got "#2"
FAIL traversals in the same (forward) direction: coalesced assert_equals: first load event must be going forward (hash) expected "#3" but got ""
